requirements.txt
pandas>=2.0
openpyxl>=3.1
python-dotenv>=1.0
mysql-connector-python>=8.0
reportlab>=4